requirements.txt
selenium
Faker
python-dotenv
